fix c/p error in Double conversion found by unit tests (yay!)
git-svn-id: https://svn.apache.org/repos/asf/velocity/tools/branches/2.x@551311 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/velocity/tools/generic/ConversionTool.java
@@ -224,7 +224,7 @@ public class ConversionTool extends LocaleConfig
             return (Double)value;
         }
         Number num = toNumber(value);
-        return new Double(num.intValue());
+        return new Double(num.doubleValue());
     }
 
     /**
